Turn two $ffa0 that are not related to hSCY back into literals
These two additions are instead moving a cursor from the vram
coordinates of one high-score row to the previous row;
since each row is three tiles tall, that is subtracting $60,
or equivalently adding $ffa0.
Without this change and if hSCY is moved, the high score table will
have names and scores drawn to the wrong location.